+++ packages/asterui/src/codeeditor.ts
@@ -1,0 +1,1 @@
+export { CodeEditor, type CodeEditorProps, type CodeEditorLanguage } from './components/CodeEditor'